requirements.txt
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
email-validator==2.3.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg==3.2.11
psycopg-binary==3.2.11
psycopg2-binary==2.9.11
python-slugify==8.0.4
SQLAlchemy==2.0.44
text-unidecode==1.3
typing_extensions==4.15.0
Werkzeug==3.1.3
WTForms==3.2.1
zipp==3.23.0
